requirements.txt
django == 2.0
django-localized-fields == 4.6a3
django-watson == 1.5.2
psycopg2 == 2.7.4
django-markdownx == 2.0.21
django-debug-toolbar == 1.9.1
django-simple-captcha == 0.5.6
django-filter == 1.1
